Add support for @custody/plugin-command-start-debugger
By

1. reporting the child PID and V8 inspector URL (if any);
2. and overriding ‘SIGUSR1’ to start the debugger on a dynamic port, to enable
  debugging multiple processes simultaneously.

This is slightly hacky, since we now report information about the process,
and even modify the process environment, for the benefit of a custody plugin
vs. custody itself. However I think this change will be sufficiently useful to
bundle this into the default probe vs. requiring
@custody/plugin-command-start-debugger users to install a secondary
probe; the modification is not very disruptive IMO; and the modification can be
disabled if users wish.